--- src/Routes/Badger.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { Route } from "react-router-dom";
-
-import Badger from "Badger/";
-import { BadgesProvider } from "providers/useBadges.js";
-
-export default function PlannerRoutes() {
-  return (
-    <Route path="/badger">
-      <BadgesProvider>
-        <Badger />
-      </BadgesProvider>
-    </Route>
-  );
-}
