--- src/Routes/Home.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { Route } from "react-router-dom";
-
-import Home from "Home/";
-
-export default function PlannerRoutes() {
-  return (
-    <Route path="/">
-      <Home />
-    </Route>
-  );
-}
